feat(assumptions) : replace Eq with Q.eq
Move _DeprecatedEq and _DeprecatedNe to deprecated.py
